refactor: Split Store god object into focused modules (#125)
Split src/store.rs (1,916 lines) into 6 smaller files:

src/store/
  mod.rs       468 lines  (Store struct, open/init, FTS, RRF)
  chunks.rs    352 lines  (chunk CRUD operations)
  notes.rs     197 lines  (note CRUD and search)
  calls.rs     220 lines  (call graph storage/queries)
  helpers.rs   245 lines  (types, embedding conversion)
src/search.rs  531 lines  (search algorithms, scoring)

Largest file reduced from 1,916 to 531 lines (3.6x reduction).

Also fixes name_match_score to preserve camelCase boundaries
when tokenizing (tokenize before lowercasing, not after).
