+++ src/lib.rs
@@ -60,6 +60,7 @@ pub mod mcp;
 pub mod nl;
 pub mod note;
 pub mod parser;
+pub mod search;
 pub mod source;
 pub mod store;
