+++ src/store/helpers.rs
@@ -1,0 +1,245 @@
+//! Store helper types and embedding conversion functions
+
+use std::collections::HashMap;
+use std::path::PathBuf;
+use thiserror::Error;
+
+use crate::embedder::Embedding;
+use crate::parser::{ChunkType, Language};
+
+// Schema version for migrations
+pub const CURRENT_SCHEMA_VERSION: i32 = 10;
+pub const MODEL_NAME: &str = "intfloat/e5-base-v2";
+
+#[derive(Error, Debug)]
+pub enum StoreError {
+    #[error("Database error: {0}")]
+    Database(#[from] sqlx::Error),
+    #[error("IO error: {0}")]
+    Io(#[from] std::io::Error),
+    #[error("Runtime error: {0}")]
+    Runtime(String),
+    #[error("Schema version mismatch: index is v{0}, cq expects v{1}. Run 'cq index --force' to rebuild.")]
+    SchemaMismatch(i32, i32),
+    #[error("Index created by newer cq version (schema v{0}). Please upgrade cq.")]
+    SchemaNewerThanCq(i32),
+    #[error(
+        "Model mismatch: index uses '{0}', current is '{1}'. Run 'cq index --force' to re-embed."
+    )]
+    ModelMismatch(String, String),
+}
+
+/// Raw row from chunks table (used by search module)
+#[derive(Clone)]
+pub struct ChunkRow {
+    pub id: String,
+    pub origin: String,
+    pub language: String,
+    pub chunk_type: String,
+    pub name: String,
+    pub signature: String,
+    pub content: String,
+    pub doc: Option<String>,
+    pub line_start: u32,
+    pub line_end: u32,
+    pub parent_id: Option<String>,
+}
+
+/// Chunk metadata returned from search results
+///
+/// Contains all chunk information except the embedding vector.
+#[derive(Debug, Clone)]
+pub struct ChunkSummary {
+    /// Unique identifier
+    pub id: String,
+    /// Source file path (relative to project root)
+    pub file: PathBuf,
+    /// Programming language
+    pub language: Language,
+    /// Type of code element
+    pub chunk_type: ChunkType,
+    /// Name of the function/class/etc.
+    pub name: String,
+    /// Function signature or declaration
+    pub signature: String,
+    /// Full source code
+    pub content: String,
+    /// Documentation comment if present
+    pub doc: Option<String>,
+    /// Starting line number (1-indexed)
+    pub line_start: u32,
+    /// Ending line number (1-indexed)
+    pub line_end: u32,
+}
+
+impl From<ChunkRow> for ChunkSummary {
+    fn from(row: ChunkRow) -> Self {
+        ChunkSummary {
+            id: row.id,
+            file: PathBuf::from(row.origin),
+            language: row.language.parse().unwrap_or(Language::Rust),
+            chunk_type: row.chunk_type.parse().unwrap_or(ChunkType::Function),
+            name: row.name,
+            signature: row.signature,
+            content: row.content,
+            doc: row.doc,
+            line_start: row.line_start,
+            line_end: row.line_end,
+        }
+    }
+}
+
+/// A search result with similarity score
+#[derive(Debug)]
+pub struct SearchResult {
+    /// The matching chunk
+    pub chunk: ChunkSummary,
+    /// Similarity score (0.0 to 1.0, higher is better)
+    pub score: f32,
+}
+
+/// Caller information from the full call graph
+///
+/// Unlike ChunkSummary, this doesn't require a chunk to exist -
+/// it captures callers from large functions that exceed chunk size limits.
+#[derive(Debug, Clone)]
+pub struct CallerInfo {
+    /// Function name
+    pub name: String,
+    /// Source file path
+    pub file: PathBuf,
+    /// Line where function starts
+    pub line: u32,
+}
+
+/// Note metadata returned from search results
+#[derive(Debug, Clone)]
+pub struct NoteSummary {
+    /// Unique identifier
+    pub id: String,
+    /// Note content
+    pub text: String,
+    /// Sentiment: -1.0 to +1.0
+    pub sentiment: f32,
+    /// Mentioned code paths/functions
+    pub mentions: Vec<String>,
+}
+
+/// A note search result with similarity score
+#[derive(Debug)]
+pub struct NoteSearchResult {
+    /// The matching note
+    pub note: NoteSummary,
+    /// Similarity score (0.0 to 1.0)
+    pub score: f32,
+}
+
+/// Unified search result (code chunk or note)
+#[derive(Debug)]
+pub enum UnifiedResult {
+    Code(SearchResult),
+    Note(NoteSearchResult),
+}
+
+impl UnifiedResult {
+    /// Get the similarity score
+    pub fn score(&self) -> f32 {
+        match self {
+            UnifiedResult::Code(r) => r.score,
+            UnifiedResult::Note(r) => r.score,
+        }
+    }
+}
+
+/// Filter and scoring options for search
+///
+/// All fields are optional. Unset filters match all chunks.
+#[derive(Default)]
+pub struct SearchFilter {
+    /// Filter by programming language(s)
+    pub languages: Option<Vec<Language>>,
+    /// Filter by file path glob pattern (e.g., `src/**/*.rs`)
+    pub path_pattern: Option<String>,
+    /// Weight for name matching in hybrid search (0.0-1.0)
+    ///
+    /// 0.0 = pure embedding similarity (default)
+    /// 1.0 = pure name matching
+    /// 0.2 = recommended for balanced results
+    pub name_boost: f32,
+    /// Query text for name matching (required if name_boost > 0 or enable_rrf)
+    pub query_text: String,
+    /// Enable RRF (Reciprocal Rank Fusion) hybrid search
+    ///
+    /// When enabled, combines semantic search results with FTS5 keyword search
+    /// using the formula: score = Σ 1/(k + rank), where k=60.
+    /// This typically improves recall for identifier-heavy queries.
+    pub enable_rrf: bool,
+}
+
+/// Model metadata for index initialization
+pub struct ModelInfo {
+    pub name: String,
+    pub dimensions: u32,
+    pub version: String,
+}
+
+impl Default for ModelInfo {
+    fn default() -> Self {
+        ModelInfo {
+            name: MODEL_NAME.to_string(),
+            dimensions: 769, // 768 from model + 1 sentiment
+            version: "1.5".to_string(),
+        }
+    }
+}
+
+/// Index statistics
+#[derive(Debug)]
+pub struct IndexStats {
+    pub total_chunks: u64,
+    pub total_files: u64,
+    pub chunks_by_language: HashMap<Language, u64>,
+    pub chunks_by_type: HashMap<ChunkType, u64>,
+    pub index_size_bytes: u64,
+    pub created_at: String,
+    pub updated_at: String,
+    pub model_name: String,
+    pub schema_version: i32,
+}
+
+// ============ Embedding Serialization ============
+
+/// Convert embedding to bytes for storage
+pub fn embedding_to_bytes(embedding: &Embedding) -> Vec<u8> {
+    embedding
+        .as_slice()
+        .iter()
+        .flat_map(|f| f.to_le_bytes())
+        .collect()
+}
+
+/// Zero-copy view of embedding bytes as f32 slice (for hot paths)
+pub fn embedding_slice(bytes: &[u8]) -> Option<&[f32]> {
+    const EXPECTED_BYTES: usize = 769 * 4; // 768 model + 1 sentiment
+    if bytes.len() != EXPECTED_BYTES {
+        return None;
+    }
+    Some(bytemuck::cast_slice(bytes))
+}
+
+/// Convert embedding bytes to owned Vec (when ownership needed)
+pub fn bytes_to_embedding(bytes: &[u8]) -> Vec<f32> {
+    embedding_slice(bytes)
+        .map(|s| s.to_vec())
+        .unwrap_or_else(|| {
+            tracing::warn!(
+                "Embedding byte length mismatch: expected {}, got {} (possible corruption)",
+                769 * 4,
+                bytes.len()
+            );
+            bytes
+                .chunks_exact(4)
+                .map(|chunk| f32::from_le_bytes(chunk.try_into().expect("4 bytes")))
+                .collect()
+        })
+}
